requirements.txt
# Architecture Governance System - Python Dependencies

# Environment management
python-dotenv>=1.0.0

# Confluence integration
atlassian-python-api>=3.41.0

# HTML parsing (for Confluence content conversion)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Markdown processing
markdown>=3.5.0
markdownify>=0.11.0

# HTTP client
requests>=2.31.0

# HTML templating (for reports)
Jinja2>=3.1.0

# PowerPoint generation (for lifecycle presentation export)
python-pptx>=1.0.0

# FastAPI watcher server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
